Move more queries to registerQuery
