+++ core/src/main/java/org/elasticsearch/index/query/DisMaxQueryBuilder.java
@@ -21,9 +21,12 @@ package org.elasticsearch.index.query;
 
 import org.apache.lucene.search.DisjunctionMaxQuery;
 import org.apache.lucene.search.Query;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -39,14 +42,17 @@ import java.util.Objects;
 public class DisMaxQueryBuilder extends AbstractQueryBuilder<DisMaxQueryBuilder> {
 
     public static final String NAME = "dis_max";
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+    public static final DisMaxQueryBuilder PROTOTYPE = new DisMaxQueryBuilder();
+
+    private static final ParseField TIE_BREAKER_FIELD = new ParseField("tie_breaker");
+    private static final ParseField QUERIES_FIELD = new ParseField("queries");
 
     private final List<QueryBuilder<?>> queries = new ArrayList<>();
 
     /** Default multiplication factor for breaking ties in document scores.*/
     public static float DEFAULT_TIE_BREAKER = 0.0f;
     private float tieBreaker = DEFAULT_TIE_BREAKER;
-
-    static final DisMaxQueryBuilder PROTOTYPE = new DisMaxQueryBuilder();
 
     /**
      * Add a sub-query to this disjunction.
@@ -88,14 +94,75 @@ public class DisMaxQueryBuilder extends AbstractQueryBuilder<DisMaxQueryBuilder>
     @Override
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject(NAME);
-        builder.field(DisMaxQueryParser.TIE_BREAKER_FIELD.getPreferredName(), tieBreaker);
-        builder.startArray(DisMaxQueryParser.QUERIES_FIELD.getPreferredName());
+        builder.field(TIE_BREAKER_FIELD.getPreferredName(), tieBreaker);
+        builder.startArray(QUERIES_FIELD.getPreferredName());
         for (QueryBuilder<?> queryBuilder : queries) {
             queryBuilder.toXContent(builder, params);
         }
         builder.endArray();
         printBoostAndQueryName(builder);
         builder.endObject();
+    }
+
+    public static DisMaxQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+
+        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
+        float tieBreaker = DisMaxQueryBuilder.DEFAULT_TIE_BREAKER;
+
+        final List<QueryBuilder<?>> queries = new ArrayList<>();
+        boolean queriesFound = false;
+        String queryName = null;
+
+        String currentFieldName = null;
+        XContentParser.Token token;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (token == XContentParser.Token.START_OBJECT) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, QUERIES_FIELD)) {
+                    queriesFound = true;
+                    QueryBuilder<?> query = parseContext.parseInnerQueryBuilder();
+                    queries.add(query);
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[dis_max] query does not support [" + currentFieldName + "]");
+                }
+            } else if (token == XContentParser.Token.START_ARRAY) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, QUERIES_FIELD)) {
+                    queriesFound = true;
+                    while (token != XContentParser.Token.END_ARRAY) {
+                        QueryBuilder<?> query = parseContext.parseInnerQueryBuilder();
+                        queries.add(query);
+                        token = parser.nextToken();
+                    }
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[dis_max] query does not support [" + currentFieldName + "]");
+                }
+            } else {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
+                    boost = parser.floatValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, TIE_BREAKER_FIELD)) {
+                    tieBreaker = parser.floatValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
+                    queryName = parser.text();
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[dis_max] query does not support [" + currentFieldName + "]");
+                }
+            }
+        }
+
+        if (!queriesFound) {
+            throw new ParsingException(parser.getTokenLocation(), "[dis_max] requires 'queries' field");
+        }
+
+        DisMaxQueryBuilder disMaxQuery = new DisMaxQueryBuilder();
+        disMaxQuery.tieBreaker(tieBreaker);
+        disMaxQuery.queryName(queryName);
+        disMaxQuery.boost(boost);
+        for (QueryBuilder<?> query : queries) {
+            disMaxQuery.add(query);
+        }
+        return disMaxQuery;
     }
 
     @Override
