--- core/src/main/java/org/elasticsearch/index/query/HasChildQueryParser.java
@@ -1,123 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.index.query;
-
-import org.apache.lucene.search.join.ScoreMode;
-import org.elasticsearch.common.ParseField;
-import org.elasticsearch.common.ParsingException;
-import org.elasticsearch.common.xcontent.XContentParser;
-import org.elasticsearch.index.query.support.InnerHitBuilder;
-
-import java.io.IOException;
-import java.util.Locale;
-
-/**
- * A query parser for <tt>has_child</tt> queries.
- */
-public class HasChildQueryParser implements QueryParser<HasChildQueryBuilder> {
-
-    public static final ParseField QUERY_NAME_FIELD = new ParseField(HasChildQueryBuilder.NAME);
-    public static final ParseField QUERY_FIELD = new ParseField("query", "filter");
-    public static final ParseField TYPE_FIELD = new ParseField("type", "child_type");
-    public static final ParseField MAX_CHILDREN_FIELD = new ParseField("max_children");
-    public static final ParseField MIN_CHILDREN_FIELD = new ParseField("min_children");
-    public static final ParseField SCORE_MODE_FIELD = new ParseField("score_mode");
-    public static final ParseField INNER_HITS_FIELD = new ParseField("inner_hits");
-
-    @Override
-    public HasChildQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
-        XContentParser parser = parseContext.parser();
-        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
-        String childType = null;
-        ScoreMode scoreMode = HasChildQueryBuilder.DEFAULT_SCORE_MODE;
-        int minChildren = HasChildQueryBuilder.DEFAULT_MIN_CHILDREN;
-        int maxChildren = HasChildQueryBuilder.DEFAULT_MAX_CHILDREN;
-        String queryName = null;
-        InnerHitBuilder innerHitBuilder = null;
-        String currentFieldName = null;
-        XContentParser.Token token;
-        QueryBuilder iqb = null;
-        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
-            if (token == XContentParser.Token.FIELD_NAME) {
-                currentFieldName = parser.currentName();
-            } else if (parseContext.isDeprecatedSetting(currentFieldName)) {
-                // skip
-            } else if (token == XContentParser.Token.START_OBJECT) {
-                if (parseContext.parseFieldMatcher().match(currentFieldName, QUERY_FIELD)) {
-                    iqb = parseContext.parseInnerQueryBuilder();
-                } else if (parseContext.parseFieldMatcher().match(currentFieldName, INNER_HITS_FIELD)) {
-                    innerHitBuilder = InnerHitBuilder.fromXContent(parser, parseContext);
-                } else {
-                    throw new ParsingException(parser.getTokenLocation(), "[has_child] query does not support [" + currentFieldName + "]");
-                }
-            } else if (token.isValue()) {
-                if (parseContext.parseFieldMatcher().match(currentFieldName, TYPE_FIELD)) {
-                    childType = parser.text();
-                } else if (parseContext.parseFieldMatcher().match(currentFieldName, SCORE_MODE_FIELD)) {
-                    scoreMode = parseScoreMode(parser.text());
-                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
-                    boost = parser.floatValue();
-                } else if (parseContext.parseFieldMatcher().match(currentFieldName, MIN_CHILDREN_FIELD)) {
-                    minChildren = parser.intValue(true);
-                } else if (parseContext.parseFieldMatcher().match(currentFieldName, MAX_CHILDREN_FIELD)) {
-                    maxChildren = parser.intValue(true);
-                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
-                    queryName = parser.text();
-                } else {
-                    throw new ParsingException(parser.getTokenLocation(), "[has_child] query does not support [" + currentFieldName + "]");
-                }
-            }
-        }
-        HasChildQueryBuilder hasChildQueryBuilder = new HasChildQueryBuilder(childType, iqb, maxChildren, minChildren,
-                scoreMode, innerHitBuilder);
-        hasChildQueryBuilder.queryName(queryName);
-        hasChildQueryBuilder.boost(boost);
-        return hasChildQueryBuilder;
-    }
-
-    public static ScoreMode parseScoreMode(String scoreModeString) {
-        if ("none".equals(scoreModeString)) {
-            return ScoreMode.None;
-        } else if ("min".equals(scoreModeString)) {
-            return ScoreMode.Min;
-        } else if ("max".equals(scoreModeString)) {
-            return ScoreMode.Max;
-        } else if ("avg".equals(scoreModeString)) {
-            return ScoreMode.Avg;
-        } else if ("sum".equals(scoreModeString)) {
-            return ScoreMode.Total;
-        }
-        throw new IllegalArgumentException("No score mode for child query [" + scoreModeString + "] found");
-    }
-
-    public static String scoreModeAsString(ScoreMode scoreMode) {
-        if (scoreMode == ScoreMode.Total) {
-            // Lucene uses 'total' but 'sum' is more consistent with other elasticsearch APIs
-            return "sum";
-        } else {
-            return scoreMode.name().toLowerCase(Locale.ROOT);
-        }
-    }
-
-    @Override
-    public HasChildQueryBuilder getBuilderPrototype() {
-        return HasChildQueryBuilder.PROTOTYPE;
-    }
-}
