+++ core/src/main/java/org/elasticsearch/index/query/HasParentQueryBuilder.java
@@ -22,10 +22,13 @@ import org.apache.lucene.search.BooleanClause;
 import org.apache.lucene.search.BooleanQuery;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.join.ScoreMode;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.lucene.search.Queries;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.index.fielddata.plain.ParentChildIndexFieldData;
 import org.elasticsearch.index.mapper.DocumentMapper;
 import org.elasticsearch.index.mapper.internal.ParentFieldMapper;
@@ -42,8 +45,18 @@ import java.util.Set;
 public class HasParentQueryBuilder extends AbstractQueryBuilder<HasParentQueryBuilder> {
 
     public static final String NAME = "has_parent";
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+    public static final HasParentQueryBuilder PROTOTYPE = new HasParentQueryBuilder("", EmptyQueryBuilder.PROTOTYPE);
+
     public static final boolean DEFAULT_SCORE = false;
-    private final QueryBuilder query;
+
+    private static final ParseField QUERY_FIELD = new ParseField("query", "filter");
+    private static final ParseField SCORE_MODE_FIELD = new ParseField("score_mode").withAllDeprecated("score");
+    private static final ParseField TYPE_FIELD = new ParseField("parent_type", "type");
+    private static final ParseField SCORE_FIELD = new ParseField("score");
+    private static final ParseField INNER_HITS_FIELD = new ParseField("inner_hits");
+
+    private final QueryBuilder<?> query;
     private final String type;
     private boolean score = DEFAULT_SCORE;
     private InnerHitBuilder innerHit;
@@ -52,7 +65,7 @@ public class HasParentQueryBuilder extends AbstractQueryBuilder<HasParentQueryBu
      * @param type  The parent type
      * @param query The query that will be matched with parent documents
      */
-    public HasParentQueryBuilder(String type, QueryBuilder query) {
+    public HasParentQueryBuilder(String type, QueryBuilder<?> query) {
         if (type == null) {
             throw new IllegalArgumentException("[" + NAME + "] requires 'parent_type' field");
         }
@@ -63,7 +76,7 @@ public class HasParentQueryBuilder extends AbstractQueryBuilder<HasParentQueryBu
         this.query = query;
     }
 
-    public HasParentQueryBuilder(String type, QueryBuilder query, boolean score, InnerHitBuilder innerHit) {
+    public HasParentQueryBuilder(String type, QueryBuilder<?> query, boolean score, InnerHitBuilder innerHit) {
         this(type, query);
         this.score = score;
         this.innerHit = innerHit;
@@ -94,7 +107,7 @@ public class HasParentQueryBuilder extends AbstractQueryBuilder<HasParentQueryBu
     /**
      * Returns the query to execute.
      */
-    public QueryBuilder query() {
+    public QueryBuilder<?> query() {
         return query;
     }
 
@@ -185,15 +198,64 @@ public class HasParentQueryBuilder extends AbstractQueryBuilder<HasParentQueryBu
     @Override
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject(NAME);
-        builder.field(HasParentQueryParser.QUERY_FIELD.getPreferredName());
+        builder.field(QUERY_FIELD.getPreferredName());
         query.toXContent(builder, params);
-        builder.field(HasParentQueryParser.TYPE_FIELD.getPreferredName(), type);
-        builder.field(HasParentQueryParser.SCORE_FIELD.getPreferredName(), score);
+        builder.field(TYPE_FIELD.getPreferredName(), type);
+        builder.field(SCORE_FIELD.getPreferredName(), score);
         printBoostAndQueryName(builder);
         if (innerHit != null) {
-            builder.field(HasParentQueryParser.INNER_HITS_FIELD.getPreferredName(), innerHit, params);
+            builder.field(INNER_HITS_FIELD.getPreferredName(), innerHit, params);
         }
         builder.endObject();
+    }
+
+    public static HasParentQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
+        String parentType = null;
+        boolean score = HasParentQueryBuilder.DEFAULT_SCORE;
+        String queryName = null;
+        InnerHitBuilder innerHits = null;
+
+        String currentFieldName = null;
+        XContentParser.Token token;
+        QueryBuilder<?> iqb = null;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (token == XContentParser.Token.START_OBJECT) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, QUERY_FIELD)) {
+                    iqb = parseContext.parseInnerQueryBuilder();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, INNER_HITS_FIELD)) {
+                    innerHits = InnerHitBuilder.fromXContent(parser, parseContext);
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[has_parent] query does not support [" + currentFieldName + "]");
+                }
+            } else if (token.isValue()) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, TYPE_FIELD)) {
+                    parentType = parser.text();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, SCORE_MODE_FIELD)) {
+                    String scoreModeValue = parser.text();
+                    if ("score".equals(scoreModeValue)) {
+                        score = true;
+                    } else if ("none".equals(scoreModeValue)) {
+                        score = false;
+                    } else {
+                        throw new ParsingException(parser.getTokenLocation(), "[has_parent] query does not support [" +
+                                scoreModeValue + "] as an option for score_mode");
+                    }
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, SCORE_FIELD)) {
+                    score = parser.booleanValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
+                    boost = parser.floatValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
+                    queryName = parser.text();
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[has_parent] query does not support [" + currentFieldName + "]");
+                }
+            }
+        }
+        return new HasParentQueryBuilder(parentType, iqb, score, innerHits).queryName(queryName).boost(boost);
     }
 
     @Override
@@ -241,7 +303,7 @@ public class HasParentQueryBuilder extends AbstractQueryBuilder<HasParentQueryBu
 
     @Override
     protected QueryBuilder<?> doRewrite(QueryRewriteContext queryShardContext) throws IOException {
-        QueryBuilder rewrite = query.rewrite(queryShardContext);
+        QueryBuilder<?> rewrite = query.rewrite(queryShardContext);
         if (rewrite != query) {
             HasParentQueryBuilder hasParentQueryBuilder = new HasParentQueryBuilder(type, rewrite);
             hasParentQueryBuilder.score(score);
