+++ core/src/main/java/org/elasticsearch/index/query/IdsQueryBuilder.java
@@ -22,18 +22,23 @@ package org.elasticsearch.index.query;
 import org.apache.lucene.queries.TermsQuery;
 import org.apache.lucene.search.Query;
 import org.elasticsearch.cluster.metadata.MetaData;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.lucene.search.Queries;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.index.mapper.Uid;
 import org.elasticsearch.index.mapper.internal.UidFieldMapper;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 
@@ -43,12 +48,16 @@ import java.util.Set;
 public class IdsQueryBuilder extends AbstractQueryBuilder<IdsQueryBuilder> {
 
     public static final String NAME = "ids";
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+    public static final IdsQueryBuilder PROTOTYPE = new IdsQueryBuilder();
+
+    private static final ParseField TYPE_FIELD = new ParseField("type", "types", "_type");
+    private static final ParseField VALUES_FIELD = new ParseField("values");
 
     private final Set<String> ids = new HashSet<>();
 
     private final String[] types;
 
-    static final IdsQueryBuilder PROTOTYPE = new IdsQueryBuilder();
 
     /**
      * Creates a new IdsQueryBuilder without providing the types of the documents to look for
@@ -95,8 +104,8 @@ public class IdsQueryBuilder extends AbstractQueryBuilder<IdsQueryBuilder> {
     @Override
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject(NAME);
-        builder.array(IdsQueryParser.TYPE_FIELD.getPreferredName(), types);
-        builder.startArray(IdsQueryParser.VALUES_FIELD.getPreferredName());
+        builder.array(TYPE_FIELD.getPreferredName(), types);
+        builder.startArray(VALUES_FIELD.getPreferredName());
         for (String value : ids) {
             builder.value(value);
         }
@@ -104,6 +113,74 @@ public class IdsQueryBuilder extends AbstractQueryBuilder<IdsQueryBuilder> {
         printBoostAndQueryName(builder);
         builder.endObject();
     }
+
+    public static IdsQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+        List<String> ids = new ArrayList<>();
+        List<String> types = new ArrayList<>();
+        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
+        String queryName = null;
+
+        String currentFieldName = null;
+        XContentParser.Token token;
+        boolean idsProvided = false;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (token == XContentParser.Token.START_ARRAY) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, VALUES_FIELD)) {
+                    idsProvided = true;
+                    while ((token = parser.nextToken()) != XContentParser.Token.END_ARRAY) {
+                        if ((token == XContentParser.Token.VALUE_STRING) ||
+                                (token == XContentParser.Token.VALUE_NUMBER)) {
+                            String id = parser.textOrNull();
+                            if (id == null) {
+                                throw new ParsingException(parser.getTokenLocation(), "No value specified for term filter");
+                            }
+                            ids.add(id);
+                        } else {
+                            throw new ParsingException(parser.getTokenLocation(),
+                                    "Illegal value for id, expecting a string or number, got: " + token);
+                        }
+                    }
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, TYPE_FIELD)) {
+                    while ((token = parser.nextToken()) != XContentParser.Token.END_ARRAY) {
+                        String value = parser.textOrNull();
+                        if (value == null) {
+                            throw new ParsingException(parser.getTokenLocation(), "No type specified for term filter");
+                        }
+                        types.add(value);
+                    }
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[" + IdsQueryBuilder.NAME +
+                            "] query does not support [" + currentFieldName + "]");
+                }
+            } else if (token.isValue()) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, TYPE_FIELD)) {
+                    types = Collections.singletonList(parser.text());
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
+                    boost = parser.floatValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
+                    queryName = parser.text();
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[" + IdsQueryBuilder.NAME +
+                            "] query does not support [" + currentFieldName + "]");
+                }
+            } else {
+                throw new ParsingException(parser.getTokenLocation(), "[" + IdsQueryBuilder.NAME +
+                        "] unknown token [" + token + "] after [" + currentFieldName + "]");
+            }
+        }
+        if (!idsProvided) {
+            throw new ParsingException(parser.getTokenLocation(), "[" + IdsQueryBuilder.NAME + "] query, no ids values provided");
+        }
+
+        IdsQueryBuilder query = new IdsQueryBuilder(types.toArray(new String[types.size()]));
+        query.addIds(ids.toArray(new String[ids.size()]));
+        query.boost(boost).queryName(queryName);
+        return query;
+    }
+
 
     @Override
     public String getWriteableName() {
