+++ core/src/main/java/org/elasticsearch/index/query/TermQueryBuilder.java
@@ -22,7 +22,10 @@ package org.elasticsearch.index.query;
 import org.apache.lucene.index.Term;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.TermQuery;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.lucene.BytesRefs;
+import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.index.mapper.MappedFieldType;
 
 import java.io.IOException;
@@ -33,7 +36,11 @@ import java.io.IOException;
 public class TermQueryBuilder extends BaseTermQueryBuilder<TermQueryBuilder> {
 
     public static final String NAME = "term";
-    static final TermQueryBuilder PROTOTYPE = new TermQueryBuilder("name", "value");
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+    public static final TermQueryBuilder PROTOTYPE = new TermQueryBuilder("name", "value");
+
+    private static final ParseField TERM_FIELD = new ParseField("term");
+    private static final ParseField VALUE_FIELD = new ParseField("value");
 
     /** @see BaseTermQueryBuilder#BaseTermQueryBuilder(String, String) */
     public TermQueryBuilder(String fieldName, String value) {
@@ -68,6 +75,65 @@ public class TermQueryBuilder extends BaseTermQueryBuilder<TermQueryBuilder> {
     /** @see BaseTermQueryBuilder#BaseTermQueryBuilder(String, Object) */
     public TermQueryBuilder(String fieldName, Object value) {
         super(fieldName, value);
+    }
+
+    public static TermQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+
+        String queryName = null;
+        String fieldName = null;
+        Object value = null;
+        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
+        String currentFieldName = null;
+        XContentParser.Token token;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (parseContext.isDeprecatedSetting(currentFieldName)) {
+                // skip
+            } else if (token == XContentParser.Token.START_OBJECT) {
+                // also support a format of "term" : {"field_name" : { ... }}
+                if (fieldName != null) {
+                    throw new ParsingException(parser.getTokenLocation(),
+                            "[term] query does not support different field names, use [bool] query instead");
+                }
+                fieldName = currentFieldName;
+                while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+                    if (token == XContentParser.Token.FIELD_NAME) {
+                        currentFieldName = parser.currentName();
+                    } else {
+                        if (parseContext.parseFieldMatcher().match(currentFieldName, TERM_FIELD)) {
+                            value = parser.objectBytes();
+                        } else if (parseContext.parseFieldMatcher().match(currentFieldName, VALUE_FIELD)) {
+                            value = parser.objectBytes();
+                        } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
+                            queryName = parser.text();
+                        } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
+                            boost = parser.floatValue();
+                        } else {
+                            throw new ParsingException(parser.getTokenLocation(),
+                                    "[term] query does not support [" + currentFieldName + "]");
+                        }
+                    }
+                }
+            } else if (token.isValue()) {
+                if (fieldName != null) {
+                    throw new ParsingException(parser.getTokenLocation(),
+                            "[term] query does not support different field names, use [bool] query instead");
+                }
+                fieldName = currentFieldName;
+                value = parser.objectBytes();
+            } else if (token == XContentParser.Token.START_ARRAY) {
+                throw new ParsingException(parser.getTokenLocation(), "[term] query does not support array of values");
+            }
+        }
+
+        TermQueryBuilder termQuery = new TermQueryBuilder(fieldName, value);
+        termQuery.boost(boost);
+        if (queryName != null) {
+            termQuery.queryName(queryName);
+        }
+        return termQuery;
     }
 
     @Override
